Threshold Function
Changed the thresholding operation into a function. Let's me set the stringency for each run. This'll be good once i have the main body set up so that that can be one of the input parameters. Maybe have it set so it takes in a stringency string and have that choose what the high and low end thresholds are. For now though the main body just has it run at -0.9 <= p <= 0.9