Multithread Fix - 08.06.17
Got geneArrays function to work by fixing some parameter problems. Parallel might be working now due to change in start method. Still waiting to see if it runs to completion or hangs. Needed to add main body at the bottom in order to set the start method, but this should make the later I/O settings easier to implement