Mutual Information Function - 08.06.17
Had been using spearman correlation for corr matrix calculation. Now I can switch to using MI, which should perform better. Adapted from a script on stack exchange, with changes made for our specific data and for the proper number of bins in the histogram function. In main body, this is commented out. Can easily be switch with correlation function being used. DO NOT USE BOTH. Also, when making this a program later on, allow user to specify which measure is preferred.