Typo Fixes - 08.06.17
Fixed Typos that I caught on the mac. Strange error occurs with the MI_calc function in parallel so for now I'm just going to run the corr version so I can get some sort of output. It still takes a long time to run that though, so I may have to look into figuring out GPU processing. That's a later thing though